+++ Plutarch/Builtin/Value.hs
@@ -11,13 +11,14 @@ module Plutarch.Builtin.Value (
   pvalueData,
   punValueData,
   pscaleValue,
+  ppolicies,
 ) where
 
 import GHC.Generics (Generic)
 import Generics.SOP qualified as SOP
 import Plutarch.Builtin.Bool (PBool)
 import Plutarch.Builtin.ByteString (PByteString)
-import Plutarch.Builtin.Data (PData)
+import Plutarch.Builtin.Data (PBuiltinList, PData)
 import Plutarch.Builtin.Integer (PInteger)
 import Plutarch.Builtin.Opaque (POpaque)
 import Plutarch.Internal.Term (S, Term, punsafeBuiltin, (:-->))
@@ -165,3 +166,15 @@ pscaleValue ::
   forall (s :: S).
   Term s (PInteger :--> PBuiltinValue :--> PBuiltinValue)
 pscaleValue = punsafeBuiltin PLC.ScaleValue
+
+{- | Returns a list of all currency symbols in a 'PBuiltinValue' in ascending
+order. Based on the
+[CIP-168](https://github.com/fallen-icarus/CIPs/blob/value-builtins-ext/CIP-0168/README.md)
+function of the same name.
+
+@since 1.15.0
+-}
+ppolicies ::
+  forall (s :: S).
+  Term s (PBuiltinValue :--> PBuiltinList PByteString)
+ppolicies = punsafeBuiltin PLC.Policies